requirements_export.txt
pandas>=2.0.0
pyarrow>=12.0.0
psycopg2-binary>=2.9.0
